Some clean ups and small improvements
Removed the need for commons-lang entirely by using Java's own time API
to keep track of the idle duration.

Use a service listener instead of DM to let the FW do the filtering of
ignored services on our behalf. Shaves off a couple of lines of code.

Use non-blocking collections were appropriate.

Allow the timeout and polling interval also be provided by the
framework, instead of only as system property.